third_party/blink/web_tests/platform/mac-mac-arm11.0/virtual/plz-dedicated-worker/external/wpt/fetch/h1-parsing/status-code.window-expected.txt

This is a testharness.js-based test.
FAIL HTTP/1.1  (network error) assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL HTTP/1.1 BLAH (network error) assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL HTTP/1.1 0 OK  assert_equals: expected 0 but got 500
FAIL HTTP/1.1 1 OK  assert_equals: expected 1 but got 500
FAIL HTTP/1.1 99 NOT OK  assert_equals: expected 99 but got 500
FAIL HTTP/1.1 077 77  assert_equals: expected 77 but got 500
FAIL HTTP/1.1 099 HELLO  assert_equals: expected 99 but got 500
FAIL HTTP/1.1 200  assert_equals: expected 200 but got 500
FAIL HTTP/1.1 999 DOES IT MATTER  assert_equals: expected 999 but got 500
FAIL HTTP/1.1 1000 BOO (network error) assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL HTTP/1.1 0200 BOO (network error) assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL HTTP/1.1 65736 NOT 200 OR SOME SUCH (network error) assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL HTTP/1.1 131072 HI (network error) assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL HTTP/1.1 -200 TEST (network error) assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL HTTP/1.1 0xA (network error) assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL HTTP/1.1 C8 (network error) assert_unreached: Should have rejected: undefined Reached unreachable code
Harness: the test ran to completion.
